backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-phonenumber-field[phonenumbers]==7.2.0
stripe==7.4.0
paypalrestsdk==1.14.0
requests==2.31.0
python-dotenv==1.0.0
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3